Make CA more robust
