+++ ca/worker.go
@@ -1,4 +1,4 @@
-package ca2
+package ca
 
 import (
 	"context"
